02_openrouter/pyproject.toml
[project]
name = "openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
